8230168: Use ClasspathStream for FileMapInfo::create_path_array Reviewed-by: lfoltan, fparain
